requirements.txt
beautifulsoup4==4.12.2
matplotlib==3.7.1
nltk==3.8.1
pandas==2.0.2
Pillow==9.5.0
scikit_learn==1.2.2
seaborn==0.12.2
streamlit==1.24.1
textblob==0.17.1
wordcloud==1.9.2
